290 user sign up form validation (#327)
* created SignUpCitizen folder under components/Users/Auth

* changed page 1 signupcitizen to controller

* added city, state, zip validation in signupcitizen

* fixed signup citizen step 1 and step 2 and footer in termsofservices

* fixed aboutyourself

* fixed bugs for signupcitizen page

* cleaned up css and code, ready for PR

* Update client/src/components/Users/Auth/SignUpCitizen/validation-rules.ts

No longer need this :)


* Update client/src/components/Users/Auth/SignUpCitizen/SignUpCitizen.tsx


---------
