+++ client/src/components/Users/Auth/SignUpCitizen/styles.ts
@@ -1,0 +1,48 @@
+import { Theme } from '@mui/material';
+import makeStyles from '@mui/styles/makeStyles';
+import { placeholderImg } from '../../../../assets/temp';
+
+export const useStyles = makeStyles((theme: Theme) => ({
+  sideImg: {
+    backgroundImage: `url("${placeholderImg}")`,
+    backgroundSize: 'cover',
+    backgroundPosition: 'center center',
+    backgroundRepeat: 'no-repeat',
+    borderTopRightRadius: '15px',
+    borderBottomRightRadius: '15px',
+  },
+  signUpContainer: {
+    margin: theme.spacing(5),
+  },
+  button: {
+    borderRadius: 0,
+    height: 44,
+    textTransform: 'none',
+    backgroundColor: '#C4C4C4',
+    color: 'white',
+  },
+  header: {
+    fontWeight: 'bold',
+    paddingBottom: '40px',
+  },
+  input: {
+    height: 44,
+    border: '1px solid #C4C4C4',
+    borderRadius: 10,
+    boxSizing: 'border-box',
+    padding: theme.spacing(1),
+    paddingLeft: theme.spacing(2),
+    paddingRight: theme.spacing(2),
+    fontSize: 14,
+    marginBottom: 20,
+  },
+  label: {
+    color: '#000000',
+    fontWeight: 'bold',
+    textAlign: 'left',
+  },
+  chip: {
+    boxShadow: '0px 1px 2px rgba(0, 0, 0, 0.25)',
+    height: 44,
+  },
+}));
